+++ catalog/app/components/Form/index.ts
@@ -1,4 +1,4 @@
 // Form fields adapted for use with "@material-ui" and "final-form"
 
-export { default as Field } from './Field'
 export { default as Checkbox } from './Checkbox'
+export { default as TextField } from './TextField'
