+++ catalog/app/containers/Admin/Form.tsx
@@ -8,7 +8,7 @@ export interface FieldProps {
   meta: RF.FieldMetaState<string>
 }
 
-// TODO: re-use components/Form/Field
+// TODO: re-use components/Form/TextField
 export function Field({
   input,
   meta,
