+++ catalog/app/containers/Bucket/PackageDialog/PackageDialog.tsx
@@ -197,7 +197,7 @@ const useFieldInputStyles = M.makeStyles({
   },
 })
 
-// TODO: re-use components/Form/Field
+// TODO: re-use components/Form/TextField
 export function Field({
   error,
   helperText,
